+++ alor_client/src/models/bodyrequest_orders_actions_stop.rs
@@ -27,7 +27,7 @@ pub struct BodyrequestOrdersActionsStop {
   #[serde(rename = "Instrument")]
   instrument: BodyrequestOrdersActionsStopInstrument, 
   #[serde(rename = "OrderEndUnixTime")]
-  order_end_unix_time: i32,  // 0 
+  order_end_unix_time: i64,  // 0 
   #[serde(rename = "Quantity")]
   quantity: i32,  // 1 
   #[serde(rename = "Side")]
@@ -39,7 +39,7 @@ pub struct BodyrequestOrdersActionsStop {
 }
 
 impl BodyrequestOrdersActionsStop {
-  pub fn new(instrument: BodyrequestOrdersActionsStopInstrument, order_end_unix_time: i32, quantity: i32, side: Operation, trigger_price: i32, user: BodyrequestOrdersActionsStopUser, ) -> BodyrequestOrdersActionsStop {
+  pub fn new(instrument: BodyrequestOrdersActionsStopInstrument, order_end_unix_time: i64, quantity: i32, side: Operation, trigger_price: i32, user: BodyrequestOrdersActionsStopUser, ) -> BodyrequestOrdersActionsStop {
     BodyrequestOrdersActionsStop {
       instrument: instrument,
       order_end_unix_time: order_end_unix_time,
@@ -64,16 +64,16 @@ impl BodyrequestOrdersActionsStop {
   }
 
 
-  pub fn set_order_end_unix_time(&mut self, order_end_unix_time: i32) {
+  pub fn set_order_end_unix_time(&mut self, order_end_unix_time: i64) {
     self.order_end_unix_time = order_end_unix_time;
   }
 
-  pub fn with_order_end_unix_time(mut self, order_end_unix_time: i32) -> BodyrequestOrdersActionsStop {
+  pub fn with_order_end_unix_time(mut self, order_end_unix_time: i64) -> BodyrequestOrdersActionsStop {
     self.order_end_unix_time = order_end_unix_time;
     self
   }
 
-  pub fn order_end_unix_time(&self) -> &i32 {
+  pub fn order_end_unix_time(&self) -> &i64 {
     &self.order_end_unix_time
   }
