+++ alor_client/src/models/ws_res_bars_get_and_subscribe_data.rs
@@ -34,13 +34,13 @@ pub struct WsResBarsGetAndSubscribeData {
   #[serde(rename = "open")]
   open: Decimal,  // 210.82 
   #[serde(rename = "time")]
-  time: i32,  // 1620220020 
+  time: i64,  // 1620220020 
   #[serde(rename = "volume")]
   volume: i32  // 1944 
 }
 
 impl WsResBarsGetAndSubscribeData {
-  pub fn new(close: Decimal, high: Decimal, low: Decimal, open: Decimal, time: i32, volume: i32, ) -> WsResBarsGetAndSubscribeData {
+  pub fn new(close: Decimal, high: Decimal, low: Decimal, open: Decimal, time: i64, volume: i32, ) -> WsResBarsGetAndSubscribeData {
     WsResBarsGetAndSubscribeData {
       close: close,
       high: high,
@@ -107,16 +107,16 @@ impl WsResBarsGetAndSubscribeData {
   }
 
 
-  pub fn set_time(&mut self, time: i32) {
+  pub fn set_time(&mut self, time: i64) {
     self.time = time;
   }
 
-  pub fn with_time(mut self, time: i32) -> WsResBarsGetAndSubscribeData {
+  pub fn with_time(mut self, time: i64) -> WsResBarsGetAndSubscribeData {
     self.time = time;
     self
   }
 
-  pub fn time(&self) -> &i32 {
+  pub fn time(&self) -> &i64 {
     &self.time
   }
